refactor: Finished Incubator block implementation. Moa raising behavior will be worked on in another branch. Also did some more block entity related rename refactoring.
