+++ src/main/java/com/gildedgames/aether/client/registry/AetherRenderers.java
@@ -13,10 +13,10 @@ import com.gildedgames.aether.client.renderer.entity.model.*;
 import com.gildedgames.aether.client.renderer.perks.layer.DeveloperGlowLayer;
 import com.gildedgames.aether.client.renderer.perks.layer.PlayerHaloLayer;
 import com.gildedgames.aether.client.renderer.perks.model.PlayerHaloModel;
-import com.gildedgames.aether.client.renderer.tile.AetherBlockEntityWithoutLevelRenderer;
-import com.gildedgames.aether.client.renderer.tile.ChestMimicRenderer;
-import com.gildedgames.aether.client.renderer.tile.SkyrootBedRenderer;
-import com.gildedgames.aether.client.renderer.tile.TreasureChestRenderer;
+import com.gildedgames.aether.client.renderer.blockentity.AetherBlockEntityWithoutLevelRenderer;
+import com.gildedgames.aether.client.renderer.blockentity.ChestMimicRenderer;
+import com.gildedgames.aether.client.renderer.blockentity.SkyrootBedRenderer;
+import com.gildedgames.aether.client.renderer.blockentity.TreasureChestRenderer;
 import com.gildedgames.aether.client.renderer.player.layer.EnchantedDartLayer;
 import com.gildedgames.aether.client.renderer.player.layer.GoldenDartLayer;
 import com.gildedgames.aether.client.renderer.player.layer.PoisonDartLayer;
