--- src/main/java/com/gildedgames/aether/client/renderer/tile/CustomItemStackTileEntityRenderer.java
@@ -1,38 +1,0 @@
-/*
-package com.gildedgames.aether.client.renderer.tile;
-
-import java.util.function.Supplier;
-
-import com.mojang.blaze3d.vertex.PoseStack;
-
-import net.minecraft.client.model.geom.EntityModelSet;
-import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.block.model.ItemTransforms;
-import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
-import net.minecraft.client.renderer.blockentity.BlockEntityRenderDispatcher;
-import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.entity.BlockEntity;
-
-public class CustomItemStackTileEntityRenderer implements BlockEntityRenderer {
-	//private final Supplier<? extends BlockEntity> tileEntityCreator;
-	private */
-/*final*//*
- BlockEntity tileEntity;
-	
-	public CustomItemStackTileEntityRenderer(BlockEntityRenderDispatcher p_172550_, EntityModelSet p_172551_) {
-		super(p_172550_, p_172551_);
-	}
-
-	@Override
-	public void renderByItem(ItemStack itemStackIn, ItemTransforms.TransformType transformType, PoseStack matrixStackIn, MultiBufferSource bufferIn, int combinedLightIn, int combinedOverlayIn) {
-//		BlockEntity tileEntity = this.tileEntity;
-//		if (tileEntity == null) {
-//			this.tileEntity = tileEntity = tileEntityCreator.get();
-//		}
-
-		//BlockEntityRenderDispatcher.instance.renderItem(tileEntity, matrixStackIn, bufferIn, combinedLightIn, combinedOverlayIn);
-	}
-	
-}
-*/
